WIP: Invoice payment tracking, extra work improvements, and various fixes
- Add invoice payment tracking from Fakturoid webhooks
- Enhanced extra work kanban and table views with mobile support
- Creative boost invoicing protection
- Multiple database constraints and validation improvements
- Soft delete for engagements with RLS fixes
- Proration calculation fixes

+++ src/components/engagements/CreativeBoostCreditOverview.tsx
@@ -64,7 +64,7 @@ export function CreativeBoostCreditOverview({
     e.stopPropagation();
     onUpdateSettings({
       maxCredits: parseInt(tempMaxCredits) || 0,
-      pricePerCredit: parseInt(tempPricePerCredit) || 0,
+      pricePerCredit: parseFloat(tempPricePerCredit) || 0,
     });
     setIsEditing(false);
   };
